+++ src/CHANGELOG.md
@@ -1,8 +1,9 @@
 ## | 2021/06/09 | [1.1.0](https://github.com/dunc415/Collection/issues/10)
  - CSS Implementation
  - Some Features have also been added
-    - SearchBar in ViewArtist
+    - SearchBar in ViewArtist and ViewALbum
     - Proper MenuBar
+    - Popup in AddAlbum and AddArtist screen
   
 ## | 2021/06/03 | [1.0.1](https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSP5pVKEbt76K2A5zmrRmNzEwEXHkruNkDNaA&usqp=CAU)
  - Can open Album Cover on the ViewAlbum screen